refactor: EventService 리팩토링
- EventService를 Command와 Query로 분리
- AllImagesOfContent 일급컬렉션 구현

#803
