+++ backend/emm-sale/src/main/java/com/emmsale/event/api/EventApi.java
@@ -1,6 +1,7 @@
 package com.emmsale.event.api;
 
-import com.emmsale.event.application.EventService;
+import com.emmsale.event.application.EventCommandService;
+import com.emmsale.event.application.EventQueryService;
 import com.emmsale.event.application.dto.EventDetailRequest;
 import com.emmsale.event.application.dto.EventDetailResponse;
 import com.emmsale.event.domain.EventStatus;
@@ -30,11 +31,12 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class EventApi {
 
-  private final EventService eventService;
+  private final EventQueryService eventQueryService;
+  private final EventCommandService eventCommandService;
 
   @GetMapping("/{id}")
   public ResponseEntity<EventDetailResponse> findEventById(@PathVariable final Long id) {
-    return ResponseEntity.ok(eventService.findEvent(id, LocalDate.now()));
+    return ResponseEntity.ok(eventQueryService.findEvent(id, LocalDate.now()));
   }
 
   @GetMapping
@@ -46,7 +48,8 @@ public class EventApi {
       @RequestParam(required = false) final List<EventStatus> statuses,
       @RequestParam(required = false) final String keyword) {
     return ResponseEntity.ok(
-        eventService.findEvents(category, LocalDateTime.now(), startDate, endDate, tags, statuses,
+        eventQueryService.findEvents(category, LocalDateTime.now(), startDate, endDate, tags,
+            statuses,
             keyword));
   }
 
@@ -54,7 +57,7 @@ public class EventApi {
   @ResponseStatus(HttpStatus.CREATED)
   public EventDetailResponse addEvent(@RequestPart @Valid final EventDetailRequest request,
       @RequestPart final List<MultipartFile> images) {
-    return eventService.addEvent(request, images);
+    return eventCommandService.addEvent(request, images);
   }
 
   @PutMapping(path = "/{eventId}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -62,12 +65,12 @@ public class EventApi {
   public EventDetailResponse updateEvent(@PathVariable final Long eventId,
       @RequestPart @Valid final EventDetailRequest request,
       @RequestPart final List<MultipartFile> images) {
-    return eventService.updateEvent(eventId, request, images);
+    return eventCommandService.updateEvent(eventId, request, images);
   }
 
   @DeleteMapping("/{eventId}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
   public void deleteEvent(@PathVariable final Long eventId) {
-    eventService.deleteEvent(eventId);
+    eventCommandService.deleteEvent(eventId);
   }
 }
