+++ backend/emm-sale/src/main/java/com/emmsale/event/domain/EventStatus.java
@@ -8,4 +8,5 @@ public enum EventStatus {
   IN_PROGRESS,
   UPCOMING,
   ENDED;
+
 }
