+++ backend/emm-sale/src/main/java/com/emmsale/image/domain/AllImagesOfContent.java
@@ -1,0 +1,34 @@
+package com.emmsale.image.domain;
+
+import static java.util.Comparator.comparing;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class AllImagesOfContent {
+
+  private final List<Image> images;
+
+  public String extractThumbnailImage() {
+    if (images.isEmpty()) {
+      return null;
+    }
+    return images.stream()
+        .filter(Image::isThumbnail)
+        .findFirst()
+        .orElseThrow()
+        .getName();
+  }
+
+  public List<String> extractInformationImages() {
+    return images.stream()
+        .sorted(comparing(Image::getOrder))
+        .filter(Image::isNotThumbnail)
+        .map(Image::getName)
+        .collect(Collectors.toList());
+  }
+}
